Extract schema loading in dedicatd file
